Add category backups management functionality
Add functionality to this plugin, which allows backups management on a category context to all users with the appropriate capabilities.
This commit also adds an additional plugin settings page with a database query optimization option, which greatly reduces the amount
of time it takes for the query to execute, but decreases accuracy of finding the backup files to only those backups, which have been created
on the current Moodle instance (no user uploaded files). The increased query execution speed helps on large Moodle installations.
Also added slovenian translations.